Making from_chat_id in ForwardMessage an i64

+++ src/methods/forward_message.rs
@@ -9,7 +9,7 @@ pub struct ForwardMessage {
     /// Unique identifier for the target chat.
     pub chat_id: i64,
     /// Unique identifier for the chat where the original message was sent.
-    pub from_chat_id: i32,
+    pub from_chat_id: i64,
     #[builder(default)]
     #[serde(skip_serializing_if = "Option::is_none")]
     /// Sends the message silently. Users get a notification without sound.
